Adding beam_pipeline and its test for graphdef experiments.
