--- tfx/experimental/distributed_inference/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/pyi/execution_spec_test.pyi
@@ -1,11 +1,0 @@
-# (generated with --quick)
-
-import execution_spec
-from typing import Any, Type
-
-ExecutionSpec: Type[execution_spec.ExecutionSpec]
-tf: module
-
-class ExecutionSpecTest(Any):
-    __doc__: str
-    def test_spec(self) -> None: ...
