--- tfx/experimental/distributed_inference/distributed_inference/savedmodel_experiments/beam_experiment-00000-of-00001
@@ -1,2 +1,0 @@
-[15.144583]
-[36.260166]
