+++ tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/imports/beam_pipeline.imports
@@ -1,4 +1,3 @@
 tensorflow/__init__ /usr/local/google/home/jzhunybj/Desktop/tfx/tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/imports/default.pyi
 apache_beam/__init__ /usr/local/google/home/jzhunybj/Desktop/tfx/tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/imports/default.pyi
-tensorflow/core/framework/graph_pb2 /usr/local/google/home/jzhunybj/Desktop/tfx/tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/imports/default.pyi
 execution_spec /usr/local/google/home/jzhunybj/Desktop/tfx/tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/pyi/execution_spec.pyi
