+++ tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/imports/beam_pipeline_test.imports
@@ -3,8 +3,6 @@ apache_beam/__init__ /usr/local/google/home/jzhunybj/Desktop/tfx/tfx/experimenta
 apache_beam/testing/test_pipeline /usr/local/google/home/jzhunybj/Desktop/tfx/tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/imports/default.pyi
 apache_beam/testing/util /usr/local/google/home/jzhunybj/Desktop/tfx/tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/imports/default.pyi
 create_complex_graph /usr/local/google/home/jzhunybj/Desktop/tfx/tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/pyi/create_complex_graph.pyi
-tensorflow/core/framework/graph_pb2 /usr/local/google/home/jzhunybj/Desktop/tfx/tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/imports/default.pyi
-tensorflow/core/framework/node_def_pb2 /usr/local/google/home/jzhunybj/Desktop/tfx/tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/imports/default.pyi
 execution_spec /usr/local/google/home/jzhunybj/Desktop/tfx/tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/pyi/execution_spec.pyi
 graph_partition /usr/local/google/home/jzhunybj/Desktop/tfx/tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/pyi/graph_partition.pyi
 beam_pipeline /usr/local/google/home/jzhunybj/Desktop/tfx/tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/pyi/beam_pipeline.pyi
