+++ tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/imports/execution_spec.imports
@@ -1,0 +1,1 @@
+tensorflow/__init__ /usr/local/google/home/jzhunybj/Desktop/tfx/tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/imports/default.pyi
