+++ tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/.pytype/pyi/beam_pipeline.pyi
