+++ tfx/experimental/distributed_inference/graphdef_experiments/subgraph_partitioning/testdata/remote_op_b/input_names-remote_op_a-remote_op_a_1.pbtxt
